+++ src/app/pages/site-list/site-list.component.ts
@@ -1,10 +1,12 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Observable } from 'rxjs';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
 
 import { PasswordManagerService } from '../../services/password-manager.service';
 import { editSiteDTO, ISite } from '../../models/site.model';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { DialogComponent } from '../../components/dialog/dialog.component';
 
 enum FormState {
   Add = 'Add',
@@ -30,6 +32,7 @@ export class SiteListComponent implements OnInit {
 
   constructor(
     private _snackBar: MatSnackBar,
+    private dialog: MatDialog,
     private passwordManagerService: PasswordManagerService
   ) {
     this.formState = FormState.Add;
@@ -69,13 +72,23 @@ export class SiteListComponent implements OnInit {
     this.initForm = { ...this.initForm, ...site };
   }
 
-  async deleteSite(id: string) {
-    try {
-      await this.passwordManagerService.deleteSite(id);
-      this.openSnackBar('Data Deleted Successfully', 'Ok');
-    } catch (error) {
-      console.error(error);
-    }
+  deleteSite(id: string) {
+    const dialogRef = this.dialog.open(DialogComponent, {
+      width: '250px',
+    });
+
+    dialogRef.afterClosed().subscribe(async (result) => {
+      if (!result) {
+        return;
+      }
+
+      try {
+        await this.passwordManagerService.deleteSite(id);
+        this.openSnackBar('Data Deleted Successfully', 'Ok');
+      } catch (error) {
+        console.error(error);
+      }
+    });
   }
 
   openSnackBar(message: string, action: string) {
